\name{anchoredmixtures-package}
\alias{anchoredmixtures-package}
\alias{anchoredmixtures}
\docType{package}
\title{
\packageTitle{anchoredmixtures}
}
\description{
\packageDescription{anchoredmixtures}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{anchoredmixtures}
\packageIndices{anchoredmixtures}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{anchoredmixtures}

Maintainer: \packageMaintainer{anchoredmixtures}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
